refactor: Move plot_bands/plot_transport to PlotsExt and clean up deps (#35)
