+++ Project.toml
@@ -6,7 +6,6 @@ version = "0.3.0"
 
 [deps]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
-Brillouin = "23470ee3-d0df-4052-8b1a-8cbd6363e7f0"
 CodecXz = "ba30903b-d9e8-5048-a5ec-d1f5b0d4b47b"
 Comonicon = "863f3e99-da2a-4334-8734-de3dacbe5542"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
@@ -18,21 +17,23 @@ JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 NCDatasets = "85f8d34a-cbdd-5861-8df4-14fed0d494ab"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 PseudoPotentialData = "5751a51d-ac76-4487-a056-413ecf6fbe19"
-PythonPlot = "274fc56d-3b97-40fa-a1cd-1b4a50311bf9"
 Spglib = "f761d5c5-86db-4880-b97f-9680a7cccfb5"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
-StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
 
 [weakdeps]
+Brillouin = "23470ee3-d0df-4052-8b1a-8cbd6363e7f0"
 DFTK = "acf6eb54-70d9-11e9-0013-234b7a5f5337"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+PythonPlot = "274fc56d-3b97-40fa-a1cd-1b4a50311bf9"
 RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 
 [extensions]
 BoltzTraPDFTKExt = "DFTK"
 BoltzTraPNCDatasetsExt = "NCDatasets"
+BoltzTraPPlotsExt = ["Plots", "Brillouin"]
+BoltzTraPPythonPlotExt = "PythonPlot"
 BoltzTraPRecipesBaseExt = "RecipesBase"
 
 [compat]
@@ -50,12 +51,11 @@ NCDatasets = "0.14"
 NPZ = "0.4"
 Plots = "1.41.3"
 PseudoPotentialData = "0.2.4"
-PythonPlot = "1.0.6"
+PythonPlot = "1"
+RecipesBase = "1"
 Spglib = "1"
 StaticArrays = "1"
-RecipesBase = "1"
 StatsBase = "0.34"
-StatsPlots = "0.15.8"
 julia = "1.10"
 
 [extras]
